ci: migrate npm publish to GitHub Actions OIDC trusted publishing
