+++ packages/cli/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@sensigo/realm-cli",
   "version": "0.1.0",
+  "publishConfig": {
+    "access": "public"
+  },
   "description": "Run autonomous agent workflows, host an MCP server over HTTP, and manage workflow definitions and run history from the terminal.",
   "license": "Apache-2.0",
   "repository": {
