+++ packages/mcp-server/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@sensigo/realm-mcp",
   "version": "0.2.0",
+  "publishConfig": {
+    "access": "public"
+  },
   "description": "MCP server for Realm — exposes 7 tools for AI agent connections over stdio or HTTP.",
   "license": "Apache-2.0",
   "repository": {
