+++ packages/testing/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@sensigo/realm-testing",
   "version": "0.1.0",
+  "publishConfig": {
+    "access": "public"
+  },
   "description": "Testing utilities for Realm workflows — fixtures, assertions, mocks, and in-memory store.",
   "license": "Apache-2.0",
   "repository": {
